repo-indexer/requirements.txt
gitpython
requests
sentence-transformers
